pyproject.toml
[project]
name = "strava-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Rakhat",email = "hamltov31@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "stravalib (>=2.4,<3.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi[all] (>=0.122.0,<0.123.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]

[tool.poetry]
packages = [{include = "strava_platform", from = "src"}]
package-mode = false


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.6"
black = "^25.11.0"
mypy = "^1.18.2"
pytest = "^9.0.1"
alembic = "^1.17.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
